+++ dbms/postgresql/src/main/java/is/codion/dbms/postgresql/PostgreSQLDatabase.java
@@ -93,7 +93,7 @@ final class PostgreSQLDatabase extends AbstractDatabase {
 	}
 
 	@Override
-	public String limitOffsetClause(Integer limit, Integer offset) {
+	public String limitOffsetClause(Integer limit, Integer offset, boolean ordered) {
 		return createLimitOffsetClause(limit, offset);
 	}